Fix login issue: 1. Changed Admin.adminType from enum to Integer to avoid MyBatis enum mapping issues 2. Updated AuthServiceImpl to handle Integer adminType 3. Updated JWT secret key for security

+++ library-management-system/src/main/java/com/library/mapper/BookMapper.java
@@ -13,7 +13,7 @@ import java.util.List;
 public interface BookMapper {
     Book selectByIsbn(@Param("ISBN") String ISBN);
 
-    Book selectById(@Param("id") Long id);
+    Book selectById(@Param("id") Integer id);
 
     List<Book> selectByKeyword(@Param("keyword") String keyword);
